fix: merge components
